refactor: 서비스 계층 인터페이스-구현체 분리
